streamlit_app/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.3.0
plotly>=5.17.0
requests>=2.31.0
joblib>=1.3.0
python-dotenv>=1.0.0
